// frontend/app/models/user.ts
/**
 * Generated by orval v7.6.0 🍺
 * Do not edit manually.
 * Web Dev 2025 REST API
 * REST server including authentication using JWT by Roy Canani & Urir Shiber
 * OpenAPI spec version: 1.0.0
 */

export interface User {
  /** The id of the user */
  _id: string;
  /** The name of the user */
  userName?: string;
  /** The email of the user */
  email: string;
  /** The email of the user */
  password?: string;
  image?: string;
}